add name to menu, update tests
